feat: use components
