+++ packages/frontend-core/src/components/index.ts
@@ -11,3 +11,5 @@ export { default as ChangePasswordModal } from "./ChangePasswordModal.svelte"
 export { default as ProfileModal } from "./ProfileModal.svelte"
 export { default as PasswordRepeatInput } from "./PasswordRepeatInput.svelte"
 export { default as Chatbox } from "./Chatbox/index.svelte"
+export { default as ChatNavigationPanel } from "./Chatbox/ChatNavigationPanel.svelte"
+export { default as ChatConversationPanel } from "./Chatbox/ChatConversationPanel.svelte"
